+++ dockerversion/release_unix.go
@@ -1,0 +1,5 @@
+// +build !windows
+
+package dockerversion
+
+const dockerReleasePath = "/etc/docker-release"