+++ dockerversion/release_windows.go
@@ -1,0 +1,3 @@
+package dockerversion
+
+const dockerReleasePath = "docker-release"